+++ iappslx/nodejs/fastWorker.js
@@ -46,7 +46,7 @@ const configKey = 'config';
 // Known good hashes for template sets
 const supportedHashes = {
     'bigip-fast-templates': [
-        'eba538b730f38d07d7c9258d624a40a1c70cd45e0ae2d909b94e9c5215643e49', // v1.5
+        '02b5dead0c9e5aab66d40c83c6965e937c204ba60a802d19469218eeb010b3f1', // v1.5
         '99bf347ba5556df2e8c7100a97ea4c24171e436ed9f5dc9dfb446387f29e0bfe', // v1.4
         'e7eba47ac564fdc6d5ae8ae4c5eb6de3d9d22673a55a2e928ab59c8c8e16376b', // v1.3
         '316653656cfd60a256d9b92820b2f702819523496db8ca01ae3adec3bd05f08c', // v1.2